+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryClient.java
@@ -975,6 +975,7 @@ public class BigQueryClient {
 
   public void createDataset(DatasetId datasetId, Map<String, String> metadata) {
     DatasetInfo.Builder datasetInfo = DatasetInfo.newBuilder(datasetId);
+    Optional.ofNullable(bigQuery.getOptions().getLocation()).ifPresent(datasetInfo::setLocation);
     if (metadata != null && !metadata.isEmpty()) {
       Optional.ofNullable(metadata.get("bigquery_location")).ifPresent(datasetInfo::setLocation);
       Optional.ofNullable(metadata.get("comment")).ifPresent(datasetInfo::setDescription);