+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryConfig.java
@@ -48,6 +48,10 @@ public interface BigQueryConfig {
 
   String getParentProjectId();
 
+  Optional<String> getCatalogProjectId();
+
+  Optional<String> getCatalogLocation();
+
   boolean useParentProjectForMetadataOperations();
 
   boolean isViewsEnabled();